Added a sendJSON method to the HttpTransport class
For the cases where you have a struct with custom JSON field names already setup.  Instead of converting that to a map, you can just marshal it yourself and give the raw string to sendJSON.  To keep it DRY I simply copied the bulk of the send method into the sendJSON method, and then called sendJSON from the send method.